refactor: move common types
